Began implementing interface for managing trip imports.
git-svn-id: https://subversion.assembla.com/svn/treesdb/branches/aspnetmvc@83 d913c9e1-59cf-094f-9d4b-ab1779aea92f

+++ TMD.Model/Trips/TripService.cs
@@ -11,6 +11,10 @@ namespace TMD.Model.Trips
         void Save(Trip t);
         Trip FindById(int id);
         void Remove(Trip t);
+
+        IList<Trip> FindAlreadyImportedTripsByUserId(int userId);
+        IList<Trip> FindNotYetImportedTripsByUserId(int userId);
+        Trip FindLastSavedTripNotYetImportedByUserId(int userId);
     }
 
     public static class TripService
@@ -19,6 +23,15 @@ namespace TMD.Model.Trips
 
         public static void Save(Trip t)
         {
+            if (t.IsImported)
+            {
+                throw new ApplicationException("Unable to save trip because it has already been imported.");
+            }
+            if (!t.ValidateRegardingPersistence().IsValid)
+            {
+                throw new ApplicationException("Unable to save trip due to validation failure.");
+            }
+            t.SetPrivatePropertyValue<DateTime>("LastSaved", DateTime.Now);
             m_Repository.Save(t);
         }
 
@@ -29,7 +42,26 @@ namespace TMD.Model.Trips
 
         public static void Remove(Trip t)
         {
+            if (t.IsImported)
+            {
+                throw new ApplicationException("Unable to remove trip because it has already been imported.");
+            }
             m_Repository.Remove(t);
+        }
+
+        public static IList<Trip> FindAlreadyImportedTripsByUserId(int userId)
+        {
+            return m_Repository.FindAlreadyImportedTripsByUserId(userId);
+        }
+
+        public static IList<Trip> FindNotYetImportedTripsByUserId(int userId)
+        {
+            return m_Repository.FindNotYetImportedTripsByUserId(userId);
+        }
+
+        public static Trip FindLastSavedTripNotYetImportedByUserId(int userId)
+        {
+            return m_Repository.FindLastSavedTripNotYetImportedByUserId(userId);
         }
     }
 }
